Made welcome page into own view
