+++ views/docs/docs.handlebars
@@ -42,27 +42,6 @@
 				<p class="missing-in-locale-notice">{{ localeHelper locale "docs" "missingInLocale" }}</p>
 			{{/if}}
 
-			{{#if showQuickLinks}}
-				<h1>{{ localeHelper locale "docs" "quickLinks" "header" }}</h1>
-				<div class="quick-links-grid">
-					<a href="/docs/install">
-							<svg xmlns="http://www.w3.org/2000/svg" width="36" height="36" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" class="feather feather-download"><path d="M21 15v4a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2v-4"></path><polyline points="7 10 12 15 17 10"></polyline><line x1="12" y1="15" x2="12" y2="3"></line></svg>						<div>
-							<p class="header">{{ localeHelper locale "docs" "quickLinks" "links" 0 "header" }}</p>
-							<p>{{ localeHelper locale "docs" "quickLinks" "links" 0 "caption" }}</p>
-						</div>
-						<svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" class="feather feather-chevron-right"><polyline points="9 18 15 12 9 6"></polyline></svg>
-					</a>
-
-					<a href="/docs/search">
-							<svg xmlns="http://www.w3.org/2000/svg" width="36" height="36" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" class="feather feather-search"><circle cx="11" cy="11" r="8"></circle><line x1="21" y1="21" x2="16.65" y2="16.65"></line></svg>						<div>
-							<p class="header">{{ localeHelper locale "docs" "quickLinks" "links" 1 "header" }}</p>
-							<p>{{ localeHelper locale "docs" "quickLinks" "links" 1 "caption" }}</p>
-						</div>
-						<svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" class="feather feather-chevron-right"><polyline points="9 18 15 12 9 6"></polyline></svg>
-					</a>
-				</div>
-			{{/if}}
-
 			{{{ content }}}
 		</div>
 	</div>
